Add a fallback to the en version of a brand term for Fluent strings
In Fluent each term you use in a string has to be availalbe in that
locale bundle. This is a bit of a hack around that. This change looks
for the presence of a term ID in the translated string, and if it
sees one it will add the string as a new ID to a new translation
context with the en brands.ftl loaded so that it will attempt to use
the default english terms.